+++ Frontend/src/main.jsx
@@ -1,18 +1,24 @@
 import { StrictMode } from "react";
-import ReactDOM from 'react-dom/client';
-import { BrowserRouter as Router } from 'react-router-dom';
-import App from "./App.jsx";
+import ReactDOM from "react-dom/client";
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import FindRoommatesPage from "./Pages/FindRoommatesPage.jsx";
+import Root from "./Components/root/Root.jsx";
 import "./index.css";
 import { AxiosProvider } from "./Security/axios/AxiosProvider.jsx";
 
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <Root />,
+    children: [{ path: "find-roommates", element: <FindRoommatesPage /> }],
+  },
+]);
+
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
-  
   <StrictMode>
-    <Router>
-      <AxiosProvider>
-        <App />
-      </AxiosProvider>
-    </Router>
+    <AxiosProvider>
+      <RouterProvider router={router} />
+    </AxiosProvider>
   </StrictMode>
 );
